Second deploy to main

+++ frontend/src/components/EntryDetails.js
@@ -1,17 +1,22 @@
 import "../stylesheets/EntryDetail.css";
 import { useEntryContext } from "../hooks/useEntryContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+const devAPI = process.env.REACT_APP_DEVURL;
+const productionAPI = process.env.REACT_APP_PROURL;
 
 const EntryDetail = ({ entry }) => {
   const { dispatch } = useEntryContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
-    const response = await fetch(
-      `https://mern-diary-backend.onrender.com/api/entry/${entry._id}`,
-      {
-        method: "DELETE",
-      }
-    );
+    if (!user) {
+      return;
+    }
+    const response = await fetch(`${devAPI}/api/entry/${entry._id}`, {
+      method: "DELETE",
+      headers: { Authorization: `Bearer ${user.token}` },
+    });
 
     const json = await response.json();
